time_macros::format_description::public

Enum OwnedFormatItemInner

pub(crate) enum OwnedFormatItemInner {
    Literal(Vec<u8>),
    StringLiteral(String),
    Component(Component),
    Compound(Vec<Self>),
    Optional {
        format: bool,
        item: Box<Self>,
    },
    First(Vec<Self>),
}

Variants

Literal(Vec<u8>)

StringLiteral(String)

Component(Component)

Compound(Vec<Self>)

Optional

Fields

format: bool

item: Box<Self>

First(Vec<Self>)

Implementations

impl OwnedFormatItemInner

pub(crate) fn optimize(&mut self)

const fn no_op() -> Self

fn merge_consecutive_literals(&mut self) -> bool

When there are multiple consecutive literals, they can be merged into a single literal.

As there are both UTF-8 and non-UTF-8 literals, the output is UTF-8 if and only if both literals are as well.

fn unnest_trivial_compounds(&mut self) -> bool

When a compound item only contains a single item, it can be replaced with that item.

fn unnest_nested_compounds(&mut self) -> bool

When a compound item contains another compound item, the latter can be inlined into the former.

fn unnest_first_only_one(&mut self) -> bool

When a first item only contains a single item, it can be replaced with that item.

fn unnest_nested_first(&mut self) -> bool

When a first item contains another first item, the latter can be inlined into the former.

fn only_formatting_uplift_optional(&mut self) -> bool

When formatting is enabled but parsing is not, the behavior of an optional item is known ahead of time. If it is formatted, the optional item can be replaced with its inner item. If it is not formatted, it can be replace with a no-op (that will likely be removed in a later pass).

fn only_formatting_uplift_first(&mut self) -> bool

When formatting is enabled but parsing is not, the behavior of a first item is known ahead of time. It can be replaced with its first item, as the first item will always be the one that is formatted.

fn only_formatting_eliminate_end(&mut self) -> bool

fn compound_containing_empty_string(&mut self) -> bool

When a compound item contains an empty string literal, it can be removed as it has no effect.